+++ rollup.config.js
@@ -2,8 +2,8 @@ import babel from 'rollup-plugin-babel'
 
 export default {
   entry: 'src/index.js',
-  dest: 'dist/js-data-debug.js',
-  sourceMap: 'dist/js-data-debug.js.map',
+  dest: 'dist/js-data.js',
+  sourceMap: 'dist/js-data.js.map',
   format: 'umd',
   moduleName: 'JSData',
   plugins: [